+++ packages/database/src/outbox/index.ts
@@ -6,6 +6,8 @@ export * from './models/event-queue.model.js';
 export * from './models/jobs-outbox.model.js';
 export * from './models/outbox.model.js';
 
-export * from './writer/outbox.writer.js';
+export * from './writers/outbox.writer.js';
 
 export * from './types/outbox.status.js';
+
+export * from './consumers/outbox.consumer.js';