+++ packages/database/src/test/utils/index.ts
@@ -1,7 +1,9 @@
 export * from './helpers/extended-outbox-event.helper.js';
 export * from './helpers/logger-mock.helper.js';
 export * from './helpers/outbox-writer-mock.helper.js';
-export * from './helpers/outbox-writer-mock.helper.js';
 
 export * from './repositories/outbox-extended-test.repository.js';
 export * from './repositories/outbox-test.repository.js';
+
+export * from './integration.helper.js';
+export { testDataSource, initializeTestDb, closeTestDb } from '../data-source.js';