+++ packages/errors/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.5.0
+
+### Minor Changes
+
+- Added new error type for invalid outbox consumer fetch size, and renamed BaseError to BaseApiError, and added a new BaseError not API related
+
 ## 0.4.4
 
 ### Patch Changes