+++ packages/health-check-nest/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 0.1.26
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/health-check@0.1.11
+  - @volontariapp/errors-nest@0.10.0
+  - @volontariapp/bridge@0.2.9
+  - @volontariapp/bridge-nest@0.3.12
+
 ## 0.1.25
 
 ### Patch Changes